Bug fix: sliding causes walking sound+animation
If you collapse a wall or something and stand on its parts, your rigidbody will be moving very very slightly, so it would permanently play the walking sounds and animations, which didnt seem right. The new behaviour means you can be moving slightly with no sound, but it seems less egregious.

Also cleaned up the PlayerAnimation script a bit.